working code based on tests for storyController

+++ repositories/storyRepo.go
@@ -10,9 +10,9 @@ import (
 )
 
 type StoryRepository interface {
-	FindById(ID int) (*models.Story, error)
+	FindById(ID uint64) (*models.Story, error)
 	Save(story *models.Story) error
-	DeleteById(ID int) error
+	DeleteById(ID uint64) error
 	GetStoriesByField(field, fieldId string) ([]*models.Story, error)
 }
 
@@ -26,7 +26,7 @@ func NewStoryRepo(db *pgx.Conn) *StoryRepo {
 	}
 }
 
-func (r *StoryRepo) FindById(ID int) (*models.Story, error) {
+func (r *StoryRepo) FindById(ID uint64) (*models.Story, error) {
 	var storyFromID models.Story
 	err := r.db.QueryRow(context.Background(), "select * from story where id=$d", ID).Scan(&storyFromID)
 	if err != nil {
@@ -40,7 +40,7 @@ func (r *StoryRepo) Save(story *models.Story) error {
 	return nil
 }
 
-func (r *StoryRepo) DeleteById(ID int) error {
+func (r *StoryRepo) DeleteById(ID uint64) error {
 	return nil
 }
